nailgun/requirements.txt
Babel==1.3
Jinja2==2.7
Mako==0.9.1
MarkupSafe==0.18
PyYAML==3.10
SQLAlchemy>=0.9.4
alembic>=0.6.2
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
decorator==3.4.0
fysom==1.0.11
iso8601==0.1.9
jsonschema==2.3.0
kombu==3.0.16
netaddr>=0.7.10
oslo.config>=1.2.1
# TODO(prmtl): remove upper bounds after all openstack packages will
# release versions that have the same pbr requirements.
# for the day of writing this comment there was 2 outdated packages:
# - keystonemiddleware
# - python-cinderclient
oslo.serialization>=1.0.0,<=1.5.0
ply==3.4
psycopg2==2.5.1
pycrypto==2.6.1
requests>=1.2.3
simplejson>=3.3.0
six>=1.5.2
urllib3>=1.7
web.py==0.37
wsgilog==0.3
wsgiref==0.1.2
keystonemiddleware>=1.2.0
# we might still need keystone command
python-keystoneclient>=0.11,<=1.4.0
python-novaclient>=2.17.0,<=2.24.1
python-cinderclient>=1.0.7,<=1.2.1
networkx>=1.8
pydot-ng>=1.0.0
ordereddict>=1.1
importlib>=1.0.1
